renamed wallabag_compatibility_test since it doesn't do that anymore
